// frontend/src/api/generated/models/decideModerationCaseRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * ComeSebebes API
 * API RESTful da rede social de comida ComeSebebes.
 * OpenAPI spec version: v1
 */
import type { DecideModerationCaseRequestDecision } from './decideModerationCaseRequestDecision';
import type { DecideModerationCaseRequestDecisionNote } from './decideModerationCaseRequestDecisionNote';

/**
 * Decisão administrativa de um caso de moderação.
 */
export interface DecideModerationCaseRequest {
  /**
   * Decisão do caso.
   * @minLength 1
   */
  decision: DecideModerationCaseRequestDecision;
  /**
   * Justificativa obrigatória para HIDDEN e REMOVED.
   * @minLength 0
   * @maxLength 2000
   */
  decisionNote?: DecideModerationCaseRequestDecisionNote;
}